requirements.txt
numpy==1.19.1
python-igraph=0.8.2
